refactor: extract folder inspection and tree logic out of state.rs
state.rs had grown a large amount of folder-tree internals. Split:

- state/folder_inspection.rs — FolderInspection, inspect_folder() and
  is_writable_metadata() (per-folder kind/symlink/lock classification)
- state/folder/tree.rs — tree building and traversal: build_children,
  build_parent_chain, build_tree_nodes_data, expansion (toggle/set),
  flat-index helpers, visible-folder collection, and the restored-
  expansion repopulation logic; the matching tests moved with them

FolderState (state/folder.rs) now reaches the tree helpers through the
tree submodule instead of super::, and AppState uses folder::tree.
No behavior changes; AGENTS.md paths updated.
